python/code_analyzer.py
import ast
import astroid
from typing import Dict, Any

def analyze_code(file_path: str) -> Dict[str, Any]:
    with open(file_path, 'r') as file:
        code = file.read()
    
    tree = ast.parse(code)
    
    analysis = {
        'imports': [],
        'functions': [],
        'classes': [],
        'global_variables': []
    }
    
    for node in ast.walk(tree):
        if isinstance(node, ast.Import):
            for alias in node.names:
                analysis['imports'].append(alias.name)
        elif isinstance(node, ast.ImportFrom):
            for alias in node.names:
                analysis['imports'].append(f"{node.module}.{alias.name}")
        elif isinstance(node, ast.FunctionDef):
            analysis['functions'].append(node.name)
        elif isinstance(node, ast.ClassDef):
            analysis['classes'].append(node.name)
        elif isinstance(node, ast.Name) and isinstance(node.ctx, ast.Store):
            if node.id not in analysis['global_variables']:
                analysis['global_variables'].append(node.id)
    
    return analysis

def generate_code(language: str, description: str) -> str:
    # This is a simplified example. In a real-world scenario, you might use a more sophisticated
    # code generation technique, possibly leveraging AI models for this task.
    if language.lower() == 'python':
        return f"""
# Generated code based on the description: {description}
def main():
    print("Hello, World!")
    # TODO: Implement the described functionality

if __name__ == "__main__":
    main()
"""
    else:
        return f"Code generation for {language} is not supported yet."