requirements/requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
requests==2.32.3
urllib3==2.2.2
python-dotenv==1.0.0
tree-sitter==0.20.4